Updating to use fluent-mixin-rewrite-tag-name
This allows for us to use tags as part of a dynamic configuration for
the datadog event.

This will allow for more dynamic events, as opposed to having multiple
static matches to handle errors, info, etc. for every app you want to
differentiate.

Assume that a match comes through tagged with "ddevents.myapp.info"

<match ddevents.**>
  type datadog_event
  api_key yOuraPIKeyaaAAAAaA
  tags fluentevent
  alert_type $tag_parts[2]
  aggregation_key $tag_parts[1]
  msg_title "App event: ${tag_parts[1]}"
  source_type_name "fluent-${tag_parts[1])"
</match>

This would end up being aggregated as well as source typed with the name
"myapp", and would set the level based on the second tag part- info.

This requires a bit more foresight in tagging the events - something
that conveniently can be handled easily via the rewrite-tag-name plugin.